Change 'overwrite' variable
'Overwrite' variable has been removed from the OPTION structure so it does not appear in the history_rfe. Allows new participants to be preprocessed with the same parameters as the previous ones without considering the 'overwrite' parameter (otherwise it creates a new .set for old participants if 'overwrite' is different).