Group display modifications
After trying the new group visualisation functions, I noted that for the last display (each group separated with STD, DEV and MMN on the same plot), the conditions 1 and 2 have been averaged. I also want to see DEV1 and DEV2 separately, so I modified the script so it display the 3 conditions now (DEV1 alone, DEV2 alone and mean of DEV1 and DEV2).