Added a function to test existance of combination
This function allows to test for the existance of a combination between 2 sets of parameters from different steps. Needed when we want to run a step2 set of parameters that have already been run on a step1 set of parameters but we want to run it on another step1 set of parameters (otherwise step2 considered as already existant and the script stops).