Changed neural lag computation
Added a new function bt_xcorrelation2_abrbaby.m which is the same as bt_xcorrelation2.m but with an third possible value for the variable 'polarity'. It is now possible to compute the neural lag based on the maximum absolute correlation value. The polarity variable can be 'POSITIVE', 'NEGATIVE', or 'ABSOLUTE'