Added STDD condition
STDD condition corresponds to the STD conditions for 'unbalanced' number of STD. Because we select all the STD trials, we don't need to differentiate between STD1 and STD2 (as opposed to 'balanced' number of trials). So for 'unbalanced' datasets, we now export only one STDD dataset that contains all the STD epochs.