Resolve trigger artefact issues for event detection
New function that detects events that have been detected between two HF onsets because of trigger artefact and needs to be removed. The output is a list of indices that corresponds to the events that needs to be removed.
The function is added in both ERP and FFR reref_filter functions.