Move event detection function out of reref function
A separate function has been created for events detection so that it can be adapted to each issue (erg artefact, stop before end of stimulation...) and it does not appear in the main reref script.
It creates report on the trials kept or not in a _trials_description.txt file.